+++ src/components/SignUpForm/SignUpForm.styles.ts
@@ -1,0 +1,5 @@
+import styled from "@emotion/styled"
+
+export const Wrapper = styled.div`
+    width: 500px;
+  `
