--- src/components/SignUpForm/signUpForm.tsx
@@ -1,122 +1,0 @@
-import React, {useState, useCallback} from "react";
-import {Form, Input, Checkbox, Button} from "antd";
-import styled from "@emotion/styled";
-
-export default function SignupForm() {
-  // 필요한 기능 : 로그인시 해당 페이지 접속 불가하게
-
-  const [email, setEmail] = useState("");
-  const [nickname, setNick] = useState("");
-  const [password, setPassword] = useState("");
-  const [passwordCheck, setPasswordCheck] = useState("");
-  const [passwordError, setPasswordError] = useState(false);
-
-  const onChangeEmail = useCallback((e) => {
-    setEmail(e.target.value);
-  }, []);
-
-  const onChangeNick = useCallback((e) => {
-    setNick(e.target.value);
-  }, []);
-
-  const onChangePassword = useCallback((e) => {
-    setPassword(e.target.value);
-  }, []);
-
-  const onChangePasswordCheck = useCallback(
-    (e) => {
-      setPasswordError(e.target.value !== password);
-      setPasswordCheck(e.target.value);
-    },
-    [password]
-  );
-
-  const [term, setTerm] = useState("");
-  const [termError, setTermError] = useState(false);
-  const onChangeTerm = useCallback((e) => {
-    setTerm(e.target.checked);
-    setTermError(false);
-  }, []);
-
-  const onSubmit = useCallback(() => {
-    if (password !== passwordCheck) {
-      setPasswordError(true);
-      return;
-    }
-    if (!term) {
-      setTermError(true);
-      return;
-    }
-    console.log(email, nickname, password);
-  }, [email, password, passwordCheck, term]);
-
-  const Wrapper = styled.div`
-    width: 500px;
-  `
-
-  return (
-    <Wrapper>
-      <Form onFinish={onSubmit} style={{padding: 10}}>
-        <div>
-          <label htmlFor="user-email">아이디</label>
-          <br />
-          <Input
-            name="user-email"
-            type="email"
-            value={email}
-            onChange={onChangeEmail}
-            required
-          />
-        </div>
-        <div>
-          <label htmlFor="user-nick">닉네임</label>
-          <br />
-          <Input
-            name="user-nick"
-            value={nickname}
-            onChange={onChangeNick}
-            required
-          />
-        </div>
-        <div>
-          <label htmlFor="user-password">비밀번호</label>
-          <br />
-          <Input
-            name="user-password"
-            type="password"
-            value={password}
-            onChange={onChangePassword}
-            required
-          />
-        </div>
-        <div>
-          <label htmlFor="user-password-check">비밀번호체크</label>
-          <br />
-          <Input
-            name="user-password-check"
-            type="password"
-            value={passwordCheck}
-            required
-            onChange={onChangePasswordCheck}
-          />
-          {passwordError && (
-            <div style={{color: "red"}}>비밀번호가 일치하지 않습니다.</div>
-          )}
-        </div>
-        <div>
-          <Checkbox name="user-term" onChange={onChangeTerm}>
-            동의합니다.
-          </Checkbox>
-          {termError && (
-            <div style={{color: "red"}}>약관에 동의하셔야 합니다.</div>
-          )}
-        </div>
-        <div style={{marginTop: 10}}>
-          <Button type="primary" htmlType="submit">
-            가입하기
-          </Button>
-        </div>
-      </Form>
-    </Wrapper>
-  );
-}
